FBX binary exporter sorts meshes alphabetically.
This has been implemented in order to give artists control over the 
order of exported material slots.

Background: When Unreal Engine 4 renders translucent polygons, the sort 
order of different translucent materials on the same mesh is determined 
by the order of the material slots on that mesh. This commit allows 
artists to define this order simply by renaming meshes.